refactor: toast recoil 상태 index 추가
