+++ src/hooks/useToast.ts
@@ -1,7 +1,7 @@
 import { useEffect } from 'react';
 import { useRecoilValue, useResetRecoilState } from 'recoil';
 
-import { toastState } from '../states/toast/atom';
+import { toastState } from '../states/toast';
 
 const useToast = () => {
   const toastInfo = useRecoilValue(toastState);
